pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag_startups"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "RAG-based startup ideas generation"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "pandas>=2.0.0",
    "sentence-transformers>=2.2.2",
    "transformers>=4.30.0",
    "langchain-chroma>=0.1.0",
    "langsmith>=0.0.30",
    "python-dotenv>=1.0.0",
    "backoff>=2.2.1",
    "chromadb>=0.4.0",
    "wbdata>=0.3.0",
    "requests>=2.31.0",
    "redis>=5.0.1",
    "cachetools>=5.3.2",  # For in-memory caching fallback
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black==24.2.0",
    "flake8>=6.0.0",
    "isort>=5.13.2",
    "mypy>=1.0.0",
    "pre-commit>=3.5.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-include-markdown-plugin>=4.0.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["rag_startups*"]
